Move generatePath to common/generatePath.ts
