+++ packages/next-typed-href/src/common/generatePath.ts
